+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - Upcoming changes...
+
 ## [1.42.0] - 2025-12-12
 ### Added
 - Added flag `--skip-headers` to fingerprint and scan commands.