+++ CHANGELOG.md
@@ -8,7 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Upcoming changes...
 
-## [1.41.1] - 2025-11-17
+## [1.42.0] - 2025-12-17
+### Added
+- Added support for filtering uninteresting data from the beginning of source files.
+  - When using `--skip-headers` it will skip over copyright notices, import statements, comments, etc.
+  - The `--skip-headers-limit` option specifies the maximum number of lines to skip if required.
+
+## [1.41.1] - 2025-12-16
 - Use `components` instead of `purls` for vulnerability detection when converting to CycloneDX format in Folder Scan
 
 ## [1.41.0] - 2025-11-17
@@ -755,3 +761,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.40.1]: https://github.com/scanoss/scanoss.py/compare/v1.40.0...v1.40.1
 [1.41.0]: https://github.com/scanoss/scanoss.py/compare/v1.40.1...v1.41.0
 [1.41.1]: https://github.com/scanoss/scanoss.py/compare/v1.41.0...v1.41.1
+[1.42.0]: https://github.com/scanoss/scanoss.py/compare/v1.41.1...v1.42.0